changing sorting algorithm

+++ src/orders/orders.go
@@ -14,7 +14,7 @@ type Order struct {
     OrderID int64 `json:"order_id"`
     Price float64 `json:"price"`
     Range string `json:"range"`
-    SystemID int `json:"system_id"`
+    SystemID int64 `json:"system_id"`
     ItemID int `json:"type_id"`
     VolumeRemain int `json:"volume_remain"`
     VolumeTotal int `json:"volume_total"`
